+++ src/onegov.core/onegov/core/tests/test_i18n.py
@@ -177,6 +177,15 @@ def test_add_fallback_to_tail():
     assert root._fallback._fallback is grandchild
 
 
+def test_recursive_fallback_chains():
+    fallback = NullTranslations()
+    translation = NullTranslations()
+    i18n.add_fallback_to_tail(translation, fallback)
+
+    with pytest.raises(RuntimeError):
+        i18n.add_fallback_to_tail(fallback, translation)
+
+
 def test_get_translation_bound_form():
 
     class MockTranslation(object):