+++ src/onegov.core/HISTORY.rst
@@ -4,6 +4,9 @@ Changelog
 - Caches layouts number formatting options.
   [msom]
 
+- Checks for recursive translation fallback chains.
+  [msom]
+
 0.71.3 (2018-08-16)
 ~~~~~~~~~~~~~~~~~~~